pyproject.toml
[project]
name = "project-repo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.27",
    "langchain-gradientai>=0.1.19",
    "python-dotenv>=1.1.1",
    "responses>=0.25.7",
    "typing-extensions>=4.14.1",
]
